# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/discounts": {
            "get": {
                "tags": [
                    "discounts"
                ],
                "summary": "Get list of discounts",
                "description": "Returns list of discounts",
                "operationId": "getDiscountsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DiscountResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "discounts"
                ],
                "summary": "Store new discount",
                "description": "Returns discount data",
                "operationId": "storeDiscount",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreDiscountRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DiscountResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/discounts/{id}": {
            "get": {
                "tags": [
                    "discounts"
                ],
                "summary": "Get discount information",
                "description": "Returns discount data",
                "operationId": "getDiscountById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Discount id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DiscountResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "discounts"
                ],
                "summary": "Update existing discount",
                "description": "Returns updated discount data",
                "operationId": "updateDiscount",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Discount id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateDiscountRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DiscountResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "discounts"
                ],
                "summary": "Delete existing discount",
                "description": "Deletes a record and returns no content",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Discount id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Discount": {
                "title": "Discount",
                "description": "Discount model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "Name",
                        "description": "Name",
                        "type": "string",
                        "example": "A nice discount"
                    },
                    "value": {
                        "title": "Value",
                        "description": "Value",
                        "type": "number"
                    },
                    "start_date": {
                        "title": "start_date",
                        "description": "start_date",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "end_date": {
                        "title": "end_date",
                        "description": "end_date of event",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "quantity": {
                        "title": "quantity",
                        "description": "quantity of discount",
                        "type": "integer",
                        "example": 1
                    },
                    "description": {
                        "title": "Description",
                        "description": "Description",
                        "type": "string",
                        "example": "A nice description"
                    }
                },
                "type": "object"
            },
...
            "DiscountResource": {
                "title": "DiscountResource",
                "description": "Discount resource",
                "properties": {
                    "discount": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Discount"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "DiscountResource"
                }
            },
...
            "StoreDiscountRequest": {
                "title": "Store Discount request",
                "description": "Store Discount request body data",
                "required": [
                    "name, value"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of the new discount",
                        "type": "string",
                        "example": "A nice discount"
                    },
                    "value": {
                        "title": "value",
                        "description": "Value of the new discount",
                        "type": "number",
                        "format": "float",
                        "example": "10"
                    },
                    "start_date": {
                        "title": "start_date",
                        "description": "Start date of the new discount",
                        "type": "string",
                        "format": "date-time",
                        "example": "2020-01-27 17:50:45"
                    },
                    "end_date": {
                        "title": "end_date",
                        "description": "End date of the new discount",
                        "type": "string",
                        "format": "date-time",
                        "example": "2020-01-27 17:50:45"
                    },
                    "description": {
                        "title": "description",
                        "description": "Description of the new discount",
                        "type": "string",
                        "example": "A nice description"
                    }
                },
                "type": "object"
            },
...
            "UpdateDiscountRequest": {
                "title": "Update Discount request",
                "description": "Update Discount request body data",
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of the new genre",
                        "type": "string",
                        "example": "A nice genre"
                    },
                    "description": {
                        "title": "description",
                        "description": "Description of the new genre",
                        "type": "string",
                        "example": "A nice description"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "bearer_token": []
        }
    ]
}